Fixed: box shadow; renamed files; storybook added
