--- client/src/Components/Button/index.stories.tsx
@@ -1,18 +1,0 @@
-import React from 'react'
-// also exported from '@storybook/react' if you can deal with breaking changes in 6.1
-import { Story, Meta } from '@storybook/react/types-6-0'
-
-import { Button, ButtonProps } from '.'
-
-export default {
-  title: 'Example/Button',
-  component: Button
-} as Meta
-
-const Template: Story<ButtonProps> = args => <Button {...args} />
-
-export const Primary = Template.bind({})
-Primary.args = {
-  colour: 'primary',
-  label: 'Button'
-}
